Split native Zenmap scan execution helpers
